+++ .github/workflows/CD.yml
@@ -15,6 +15,10 @@ jobs:
     steps:
       -
         uses: actions/checkout@master
+      - 
+        uses: leafo/gh-actions-lua@v10
+        with:
+          luaVersion: '5.4'
       -
         uses: actions/setup-node@v2
         with:
@@ -22,7 +26,7 @@ jobs:
           registry-url: 'https://registry.npmjs.org'
       -
         run: |
-          npm run build
+          ./cli.sh build --npm
       - 
         working-directory: ./dist
         run: |
