+++ README.md
@@ -2,21 +2,7 @@
 <a href="https://gamelly.github.io/love-engine"><img width="60%" src="https://raw.githubusercontent.com/gamelly/love-engine/refs/heads/main/assets/banner3.png"></a>
 </h1> 
 
-> Reimplementation of the love2d framework to make homebrew games to gba, nds, wii, ps2 and many other platforms!
-
-#### Consoles :joystick:
-
-- [X] [Nintendo Game Boy Advanced](https://github.com/gamelly/core-native-gba)
-- [ ] [Nintendo Game Cube](https://github.com/gamelly/core-native-wii)
-- [ ] [Nintendo Wii](https://github.com/gamelly/core-native-wii)
-- [ ] [Play Station 2](https://github.com/gamelly/core-native-ps2)
-
-#### Desktop / Others :computer:
-
- - [X] [HTML5](https://www.npmjs.com/package/@gamely/love-engine)
- - [ ] [SDL2](https://github.com/gamelly/core-native-sdl)
- - [ ] [Arduino](https://github.com/gamelly/core-native-arduino)
- - [ ] [Terminal](https://github.com/gamelly/core-native-ascii)
+> Reimplementation of the love2d framework to make homebrew games to gba, nds, wii, ps2, tic80 and many other platforms!
 
 ## Documentation
 
@@ -24,6 +10,12 @@
  - [Building using cmake](https://docs.gamely.com.br/group__native#cmake)
  - [Supporting Love2D from scratch](https://docs.gamely.com.br/group__manual)
 
+
+### build
+
+```
+./cli.sh build samples/pong/main.lua --html
+```
 
 ## Configuration
