+++ package.json
@@ -1,6 +1,6 @@
 {
     "name": "@gamely/love-engine",
-    "version": "0.0.17",
+    "version": "0.0.19",
     "author": "RodrigoDornelles",
     "license": "Apache-2.0",
     "homepage": "https://docs.gamely.com.br",
@@ -9,9 +9,6 @@
     "bugs": "https://github.com/gamelly/love-engine/issues",
     "description": "Game Engine written in 100% lua that runs in a vacuum.",
     "main": "dist/main.lua",
-    "scripts": {
-        "build": "rm -Rf dist;mkdir -p dist/dist;cp src/main.lua dist/dist;cp package.json dist;cp README.md dist"
-    },
     "keywords": [
         "game engine",
         "game",
