+++ src/event.lua
@@ -1,0 +1,7 @@
+--! @defgroup love_event love.event
+--! @{
+
+function love.event.quit()
+end
+
+--! @}
