+++ src/filesystem.lua
@@ -1,0 +1,10 @@
+--! @defgroup love_filesystem love.filesystem
+--! @{
+
+function love.filesystem.getInfo()
+end
+
+function love.filesystem.read()
+end
+
+--! @}
