--- src/index.html
@@ -1,46 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <title>Love Engine</title>
-    <meta name="viewport" content="width=device-width, initial-scale=1.0, maximum-scale=1.0, user-scalable=0">
-    <style>
-        *, *::after, *::before {
-            touch-action: none;
-            box-sizing: border-box;
-            border: 0;
-            padding: 0;
-            margin: 0;
-        }
-        body {
-            height: 99.998vh;
-            display: flex;
-            overflow: hidden;
-            justify-content: center;
-            align-items: start;
-            background-color: gray;
-        }
-        main {
-          position: relative;
-          width: 100%;
-          height: 100%;
-        }
-        video {
-          position: absolute;
-        }
-        canvas {
-          position: absolute;
-          image-rendering: pixelated;
-        }
-    </style>
-  </head>
-  <body>
-    <main>
-      <canvas id="gameCanvas"></canvas>
-    </main>
-    <script src="https://cdn.jsdelivr.net/npm/@gamely/core-native-html5@0.0.16/dist/index.min.js"></script>
-    <script type="module" src="https://cdn.jsdelivr.net/gh/gamelly/gly-engine@0.0.16/src/engine/core/html5/driver-wasmoon.js"></script>
-    <script>
-      document.addEventListener('DOMContentLoaded', async () => await gly.wasmoon('game.lua'))
-    </script>
-  </body>
-</html>
