+++ src/joystick.lua
@@ -1,0 +1,8 @@
+--! @defgroup love_joystick love.joystick
+--! @{
+
+function love.joystick.getJoysticks()
+    return {}
+end
+
+--! @}
