+++ src/mouse.lua
@@ -1,0 +1,11 @@
+--! @defgroup love_mouse love.mouse
+--! @{
+
+function love.mouse.setVisible()
+end
+
+function love.mouse.getPosition()
+    return 0, 0
+end
+
+--! @}
