+++ src/sound.lua
@@ -1,0 +1,8 @@
+--! @defgroup love_sound love.sound
+--! @{
+
+function love.sound.newSoundData()
+    return {}
+end
+
+--! @}
