+++ src/system.lua
@@ -1,0 +1,8 @@
+--! @defgroup love_system love.system
+--! @{
+
+function love.system.getOS()
+    return 'Web'
+end
+
+--! @}
