+++ src/timer.lua
@@ -1,0 +1,8 @@
+--! @defgroup love_timer love.timer
+--! @{
+
+function love.timer.getTime()
+    return engine.milis
+end
+
+--! @}
