+++ src/window.lua
@@ -1,0 +1,22 @@
+--! @defgroup love_window love.window
+--! @{
+
+function love.window.setMode()
+end
+
+function love.window.toPixels(w)
+    return w
+end
+
+function love.window.setTitle()
+end
+
+function love.window.setIcon()
+end
+
+function love.window.getDesktopDimensions()
+    return engine.width, engine.height
+end
+
+
+--! @}
